StockCounter - clean up code and logic, place methods in appropriate classes
